Add `pyproject.toml` file.
The `pyproject.toml` file enables installing fedbiomed as a python
package using `pip install`. It may also be used to configure some
external tools (such as linters, formatters, etc.).
This first version of the file provides with all the basic metadata
for fedbiomed, as well as with requirements, adapted from those of
the current conda environments (with some skimming and organizing).