Add `init_optimizer` to `BaseTrainingPlan`.
Note: the signature at the `BaseTrainingPlan` level is more restrictive
than that of its children, due to the current effort in preserving the
framework-specific legacy behaviors. In the future, either the base
signature should be made more flexible, or legacy behaviors should be
deprecated in favor of using the declearn-provided `Optimizer`.